Fixed type of E_VERSION_CURRENT (u8 -> u32)

+++ crates/elf/src/lib.rs
@@ -242,7 +242,7 @@ pub mod header {
 
 			pub mod version {
 				/// Field `ei_version`: Original and current version.
-				pub const EI_VERSION_CURRENT: u8 = 1;
+				pub const EI_VERSION_CURRENT: u32 = 1;
 			}
 
 			pub mod osabi {
